requirements.txt
torch
torchvision
pytorch-lightning
hydra-core
omegaconf
hydra_colorlog
numpy

# development
black
flake8
pytest
